requirements.txt
# Ember Protocol — Python Dependencies
# Install: pip install -r requirements.txt

# Backend
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.6.0

# Development & Testing
pytest>=8.0.0
Pillow>=10.0.0
